refactor(user-settings): объединены CheckboxOption и SwitchOption в один компонент
