+++ frontend/src/widgets/user-settings/ui/SettingsSidebar.tsx
@@ -21,7 +21,7 @@ export const SettingsSidebar = ({
             key={item.key}
             className={cn(
               "justify-start lg:text-lg text-[16px] py-5",
-              activeTab === item.key && "dark:bg-accent",
+              activeTab === item.key && "bg-accent",
             )}
           >
             {item.text}
